Implement resolver system for marketplace
